Create Lagged Variable and Large Bank
Here is the code for the lagged variable and the large bank. The only problem is, total assets was not included in the dataframe in the current notebook.  You need to include 'Total Assets' back into the dataframe for the large bank variable. So, before you put it into your code, make sure to go to where you define small_Quaterized_dfs and include 'Total Assets' in it. The other issue is a few lines down, we then drop all columns that have more than 500 NA values, which then drops 'Total Assets". It looks like there are 169914 NA values in Total Assets, so I do not know if we want to drop all of them, or take the mean of total assets and fill the NA values in with that?