Downloading the final df dataframe as csv
I made this notebook so we can run it once and download the df dataframe we are using as a csv. Then, we will not have to wait to run all of the code we made to create the dataframe every time we want to use it, we can just upload the df dataframe at the beginning and then you're ready to go. I was going to just upload the df csv file here, but it is too big to upload.